Feat/billable metric setup (#262)
* feat: split billableMetricSetup into small files

* refactor: use one set of metricData CRUD to handle 2 types of metric

* fix: discount code lost focus after typing

- further split billable metricsData component

* style: styling metric setup

* feat: implement graduationSetup
