+++ src/components/plan/detail/advancedConfig.tsx
@@ -1,53 +1,21 @@
+import { currencyDecimalValidate, isValidMap, showAmount } from '@/helpers'
+import { CURRENCY, IBillableMetrics, IPlan, PlanType } from '@/shared.types'
+import { FormatPainterOutlined } from '@ant-design/icons'
 import {
-  currencyDecimalValidate,
-  isValidMap,
-  randomString,
-  showAmount
-} from '@/helpers'
-import {
-  CURRENCY,
-  IBillableMetrics,
-  IPlan,
-  MetricChargeType,
-  MetricGraduatedAmount,
-  MetricLimits,
-  MetricMeteredCharge,
-  MetricType,
-  PlanType
-} from '@/shared.types'
-import {
-  DownOutlined,
-  FormatPainterOutlined,
-  MinusOutlined,
-  PlusOutlined
-} from '@ant-design/icons'
-import {
-  Badge,
   Button,
-  Col,
   Collapse,
   CollapseProps,
-  Dropdown,
-  Empty,
   Form,
   FormInstance,
   Input,
   InputNumber,
   message,
-  Modal,
-  Row,
   Select,
-  Space,
-  Tooltip,
-  Typography
+  Switch,
+  Tooltip
 } from 'antd'
-import update from 'immutability-helper'
-
-import { METRIC_CHARGE_TYPE } from '@/constants'
-import { Switch } from 'antd'
-import { PropsWithChildren, useEffect, useState } from 'react'
-import { TIME_UNITS, TNewPlan } from '.'
-import Graduation from './icons/graduation.svg?react'
+import BillableMetricSetup from './billableMetric'
+import { TIME_UNITS } from './types'
 
 interface Props {
   enableTrialWatch: boolean
@@ -61,7 +29,6 @@ interface Props {
   trialLengthUnit: number | undefined
   setTrialLengthUnit: (val: number) => void
   saveMetricData: boolean
-  plan: IPlan | TNewPlan
 }
 
 const Index = ({
@@ -75,8 +42,7 @@ const Index = ({
   selectOnetime,
   trialLengthUnit,
   setTrialLengthUnit,
-  saveMetricData,
-  plan
+  saveMetricData
 }: Props) => {
   const onTrialLengthUnitChange = (val: number) => setTrialLengthUnit(val)
   const prettifyJSON = () => {
@@ -317,10 +283,7 @@ const Index = ({
             <Form.Item label="Trial requires bank card info" name="trialDemand">
               <Switch disabled={!enableTrialWatch || formDisabled} />
             </Form.Item>
-            <div
-              className="absolute mb-6 ml-60 text-xs text-gray-400"
-              // style={{ top: '-45px', left: '240px', width: '600px' }}
-            >
+            <div className="absolute mb-6 ml-60 text-xs text-gray-400">
               When enabled, users can only use bank card payment (no Crypto or
               wire transfer) for their first purchase.
             </div>
@@ -333,6 +296,7 @@ const Index = ({
       )
     }
   ]
+
   const billableItem: CollapseProps['items'] = [
     {
       key: 'billable',
@@ -346,7 +310,6 @@ const Index = ({
             getCurrency={getCurrency}
             form={form}
             saveMetricData={saveMetricData}
-            plan={plan}
           />
         </div>
       )
@@ -411,738 +374,3 @@ const Index = ({
 }
 
 export default Index
-
-type BillableMetricSetupProps = {
-  metricsList: IBillableMetrics[] // all the billable metrics we have created, not used for edit, but used in <Select /> for user to choose.
-  getCurrency: () => CURRENCY
-  form: FormInstance
-  saveMetricData: boolean
-  plan: IPlan | TNewPlan
-}
-
-const defaultMetricLimit = (): MetricLimits & { localId: string } => ({
-  metricId: null,
-  metricLimit: 0,
-  localId: randomString(8)
-})
-const defaultMetricMeteredCharge = (): MetricMeteredCharge & {
-  localId: string
-} => ({
-  metricId: null,
-  chargeType: MetricChargeType.STANDARD,
-  standardAmount: 0,
-  standardStartValue: 0,
-  graduatedAmounts: [],
-  localId: randomString(8)
-})
-const defaultMetricRecurringCharge = (): MetricMeteredCharge & {
-  localId: string
-} => ({
-  metricId: null,
-  chargeType: MetricChargeType.STANDARD,
-  standardAmount: 0,
-  standardStartValue: 0,
-  graduatedAmounts: [],
-  localId: randomString(8)
-})
-type MetricData = {
-  metricLimits: (MetricLimits & { localId: string })[]
-  metricMeteredCharge: (MetricMeteredCharge & { localId: string })[]
-  metricRecurringCharge: (MetricMeteredCharge & { localId: string })[]
-}
-const BillableMetricSetup = ({
-  metricsList,
-  getCurrency,
-  form,
-  saveMetricData,
-  plan
-}: BillableMetricSetupProps) => {
-  const { metricLimits, metricMeteredCharge, metricRecurringCharge } = plan
-
-  const metricLimitsLocal =
-    metricLimits == null
-      ? []
-      : metricLimits.map((m) => ({
-          ...m,
-          localId: randomString(8)
-        }))
-
-  const metricMeteredChargeLocal =
-    metricMeteredCharge == null
-      ? []
-      : metricMeteredCharge.map((m) => ({
-          ...m,
-          localId: randomString(8)
-        }))
-
-  const metricRecurringChargeLocal =
-    metricRecurringCharge == null
-      ? []
-      : metricRecurringCharge.map((m) => ({
-          ...m,
-          localId: randomString(8)
-        }))
-
-  const [metricData, setMetricData] = useState<MetricData>({
-    metricLimits: metricLimitsLocal,
-    metricMeteredCharge: metricMeteredChargeLocal,
-    metricRecurringCharge: metricRecurringChargeLocal
-  })
-
-  const [graduationSetupModalOpen, setGraduationSetupModalOpen] = useState<{
-    metricType: keyof MetricData
-    localId: string
-  } | null>(null)
-
-  const addLimitData = () => {
-    setMetricData(
-      update(metricData, { metricLimits: { $push: [defaultMetricLimit()] } })
-    )
-  }
-
-  const removeLimitData = (localId: string) => {
-    const idx = metricData.metricLimits.findIndex((m) => m.localId == localId)
-    if (idx != -1) {
-      setMetricData(
-        update(metricData, { metricLimits: { $splice: [[idx, 1]] } })
-      )
-    }
-  }
-
-  const addMetricData = (type: keyof MetricData) => {
-    switch (type) {
-      case 'metricLimits':
-        setMetricData(
-          update(metricData, {
-            [type]: { $push: [defaultMetricLimit()] }
-          })
-        )
-        break
-      case 'metricMeteredCharge':
-        setMetricData(
-          update(metricData, {
-            [type]: { $push: [defaultMetricMeteredCharge()] }
-          })
-        )
-        break
-      case 'metricRecurringCharge':
-        setMetricData(
-          update(metricData, {
-            [type]: { $push: [defaultMetricRecurringCharge()] }
-          })
-        )
-        break
-    }
-  }
-
-  const removeMetricData = (type: keyof MetricData, localId: string) => {
-    const idx = metricData[type].findIndex((m) => m.localId == localId)
-    if (idx != -1) {
-      setMetricData(update(metricData, { [type]: { $splice: [[idx, 1]] } }))
-    }
-  }
-
-  const onMetricFieldChange =
-    (
-      type: keyof MetricData,
-      localId: string,
-      field: keyof MetricLimits | keyof MetricMeteredCharge
-    ) =>
-    (val: number | null) => {
-      const idx = metricData[type].findIndex((m) => m.localId == localId)
-      if (idx != -1) {
-        setMetricData(
-          update(metricData, { [type]: { [idx]: { [field]: { $set: val } } } })
-        )
-      }
-    }
-
-  const onSaveGraduationData =
-    (type: keyof MetricData, localId: string) =>
-    (graduationData: MetricGraduatedAmount[]) => {
-      const idx = metricData[type].findIndex((m) => m.localId == localId)
-      if (idx != -1) {
-        setMetricData(
-          update(metricData, {
-            [type]: { [idx]: { graduatedAmounts: { $set: graduationData } } }
-          })
-        )
-      }
-      setGraduationSetupModalOpen(null)
-    }
-
-  const onChargeTypeSelectChange =
-    (type: keyof MetricData, localId: string) => (val: number | null) => {
-      const idx = metricData[type].findIndex((m) => m.localId == localId)
-      if (idx != -1) {
-        setMetricData(
-          update(metricData, {
-            [type]: { [idx]: { chargeType: { $set: val } } }
-          })
-        )
-      }
-    }
-
-  const onMetricIdSelectChange =
-    (type: keyof MetricData, localId: string) => (val: number | null) => {
-      const idx = metricData[type].findIndex((m) => m.localId == localId)
-      if (idx != -1) {
-        setMetricData(
-          update(metricData, {
-            [type]: { [idx]: { metricId: { $set: val } } }
-          })
-        )
-      }
-    }
-
-  useEffect(() => {
-    if (saveMetricData) {
-      form.setFieldsValue(metricData)
-    }
-  }, [saveMetricData])
-
-  return (
-    <div>
-      <Form.Item label="Billable metrics" name="metricLimits" hidden={true}>
-        <Input.TextArea rows={6} />
-      </Form.Item>
-      <Form.Item
-        label="Billable metrics"
-        name="metricMeteredCharge"
-        hidden={true}
-      >
-        <Input.TextArea rows={6} />
-      </Form.Item>
-      <Form.Item
-        label="Billable metrics"
-        name="metricRecurringCharge"
-        hidden={true}
-      >
-        <Input.TextArea rows={6} />
-      </Form.Item>
-      {metricData.metricLimits.length == 0 &&
-        metricData.metricMeteredCharge.length == 0 &&
-        metricData.metricRecurringCharge.length == 0 && (
-          <Empty
-            image={Empty.PRESENTED_IMAGE_SIMPLE}
-            description="No billable metrics"
-          />
-        )}
-      {graduationSetupModalOpen != null && (
-        <GraduationSetupModal
-          data={
-            metricData[graduationSetupModalOpen.metricType].find(
-              (m) => m.localId == graduationSetupModalOpen.localId
-            )?.graduatedAmounts
-          }
-          onCancel={() => setGraduationSetupModalOpen(null)}
-          onOK={onSaveGraduationData(
-            graduationSetupModalOpen.metricType,
-            graduationSetupModalOpen.localId
-          )}
-          getCurrency={getCurrency}
-        />
-      )}
-      {metricData.metricLimits.length > 0 && (
-        <LimitSetup
-          metricData={metricData.metricLimits}
-          metricsList={metricsList.filter(
-            (m) => m.type == MetricType.LIMIT_METERED
-          )}
-          onMetricFieldChange={onMetricFieldChange}
-          onMetricIdSelectChange={onMetricIdSelectChange}
-          // getCurrency={getCurrency}
-          addLimitData={addLimitData}
-          removeLimitData={removeLimitData}
-        />
-      )}
-      {metricData.metricMeteredCharge.length > 0 && (
-        <ChargeSetup
-          metricData={metricData.metricMeteredCharge}
-          isRecurring={false}
-          metricDataType="metricMeteredCharge"
-          metricsList={metricsList.filter(
-            (m) => m.type == MetricType.CHARGE_METERED
-          )}
-          getCurrency={getCurrency}
-          addMetricData={addMetricData}
-          removeMetricData={removeMetricData}
-          onMetricFieldChange={onMetricFieldChange}
-          onChargeTypeSelectChange={onChargeTypeSelectChange}
-          onMetricIdSelectChange={onMetricIdSelectChange}
-          setGraduationSetupModalOpen={setGraduationSetupModalOpen}
-        />
-      )}
-      {metricData.metricRecurringCharge.length > 0 && (
-        <ChargeSetup
-          metricData={metricData.metricRecurringCharge}
-          isRecurring={true}
-          metricDataType="metricRecurringCharge"
-          metricsList={metricsList.filter(
-            (m) => m.type == MetricType.CHARGE_RECURRING
-          )}
-          getCurrency={getCurrency}
-          addMetricData={addMetricData}
-          removeMetricData={removeMetricData}
-          onMetricFieldChange={onMetricFieldChange}
-          onChargeTypeSelectChange={onChargeTypeSelectChange}
-          onMetricIdSelectChange={onMetricIdSelectChange}
-          setGraduationSetupModalOpen={setGraduationSetupModalOpen}
-        />
-      )}
-      <Dropdown
-        arrow={true}
-        menu={{
-          items: [
-            {
-              label: 'Limit metered',
-              disabled: metricData.metricLimits.length > 0,
-              key: MetricType.LIMIT_METERED,
-              onClick: () => {
-                addMetricData('metricLimits')
-              }
-            },
-            {
-              label: 'Charge metered',
-              disabled: metricData.metricMeteredCharge.length > 0,
-              key: MetricType.CHARGE_METERED,
-              onClick: () => {
-                addMetricData('metricMeteredCharge')
-              }
-            },
-            {
-              label: 'Charge recurring',
-              disabled: metricData.metricRecurringCharge.length > 0,
-              key: MetricType.CHARGE_RECURRING,
-              onClick: () => {
-                addMetricData('metricRecurringCharge')
-              }
-            }
-          ]
-        }}
-      >
-        <Button icon={<PlusOutlined />} variant="outlined" color="default">
-          <Space>
-            Add billing model
-            <DownOutlined />
-          </Space>
-        </Button>
-      </Dropdown>
-      {/* &nbsp;&nbsp;&nbsp;&nbsp;
-      <Button
-        onClick={() => {
-          form.setFieldsValue(metricData)
-        }}
-      >
-        Save data
-      </Button> */}
-    </div>
-  )
-}
-
-const LimitSetup = ({
-  metricData,
-  metricsList,
-  onMetricFieldChange,
-  onMetricIdSelectChange,
-  //getCurrency,
-  addLimitData,
-  removeLimitData
-}: {
-  metricData: (MetricLimits & { localId: string })[]
-  metricsList: IBillableMetrics[]
-  onMetricFieldChange: (
-    type: keyof MetricData,
-    localId: string,
-    field: keyof MetricLimits
-  ) => (val: number | null) => void
-  onMetricIdSelectChange: (
-    type: keyof MetricData,
-    localId: string
-  ) => (val: number | null) => void
-  // getCurrency: () => CURRENCY
-  addLimitData: () => void
-  removeLimitData: (localId: string) => void
-}) => {
-  // const getMetricData
-  return (
-    <div className="my-4 rounded-md bg-gray-100 p-4">
-      <Typography.Title level={5}>Limit metered</Typography.Title>
-      <Row className="mb-2">
-        <Col span={6}>Name</Col>
-        <Col span={6}>Code</Col>
-        {/* <Col>Aggregation Type</Col>
-        <Col>Aggregation Property</Col> */}
-        <Col span={10}>Limit Value</Col>
-        <Col span={2}>
-          {' '}
-          <Button
-            icon={<PlusOutlined />}
-            size="small"
-            style={{ border: 'none' }}
-            variant="outlined"
-            onClick={addLimitData}
-            color="default"
-          />
-        </Col>
-      </Row>
-      {metricData.map((m) => (
-        <Row key={m.localId} className="my-2">
-          <Col span={6}>
-            <Select
-              style={{ width: 180 }}
-              value={m.metricId}
-              onChange={onMetricIdSelectChange('metricLimits', m.localId)}
-              options={metricsList
-                .filter((m) => m.type == MetricType.LIMIT_METERED)
-                .map((m) => ({ label: m.metricName, value: m.id }))}
-            />
-          </Col>
-          <Col span={6}></Col>
-          <Col span={10}>
-            <InputNumber
-              style={{ width: 120 }}
-              placeholder="Limit value"
-              min={0}
-              value={m.metricLimit}
-              onChange={onMetricFieldChange(
-                'metricLimits',
-                m.localId,
-                'metricLimit'
-              )}
-            />
-          </Col>
-          <Col span={2}>
-            <Button
-              icon={<MinusOutlined />}
-              size="small"
-              style={{ border: 'none' }}
-              onClick={() => removeLimitData(m.localId)}
-            />
-          </Col>
-        </Row>
-      ))}
-    </div>
-  )
-}
-
-type ChargeSetupProps = {
-  metricData: (MetricMeteredCharge & { localId: string })[]
-  metricDataType: keyof MetricData
-  metricsList: IBillableMetrics[]
-  isRecurring: boolean
-  getCurrency: () => CURRENCY
-  addMetricData: (type: keyof MetricData) => void
-  removeMetricData: (type: keyof MetricData, localId: string) => void
-  onMetricFieldChange: (
-    type: keyof MetricData,
-    localId: string,
-    field: keyof MetricMeteredCharge
-  ) => (val: number | null) => void
-  onChargeTypeSelectChange: (
-    type: keyof MetricData,
-    localId: string
-  ) => (val: number | null) => void
-  onMetricIdSelectChange: (
-    type: keyof MetricData,
-    localId: string
-  ) => (val: number | null) => void
-  setGraduationSetupModalOpen: (
-    modalOpen: { metricType: keyof MetricData; localId: string } | null
-  ) => void
-}
-const rowHeaderStyle = 'text-gray-400'
-const ChargeSetup = ({
-  metricData,
-  metricDataType,
-  metricsList,
-  isRecurring,
-  getCurrency,
-  addMetricData,
-  removeMetricData,
-  onMetricFieldChange,
-  onChargeTypeSelectChange,
-  onMetricIdSelectChange,
-  setGraduationSetupModalOpen
-}: ChargeSetupProps) => {
-  return (
-    <div className="my-4 rounded-md bg-gray-100 p-4">
-      <Typography.Title level={5}>
-        Charge metered{isRecurring ? ' (recurring)' : ''}
-      </Typography.Title>
-      <Row>
-        <Col span={5} className={rowHeaderStyle}>
-          Name
-        </Col>
-        <Col span={5} className={rowHeaderStyle}>
-          {' '}
-        </Col>
-        {/* <Col>Aggregation Type</Col>
-        <Col>Aggregation Property</Col> */}
-        <Col span={6} className={rowHeaderStyle}>
-          Price
-        </Col>
-        <Col span={6} className={rowHeaderStyle}>
-          Start value
-        </Col>
-        <Col span={2}>
-          {' '}
-          <Button
-            icon={<PlusOutlined />}
-            size="small"
-            style={{ border: 'none' }}
-            variant="outlined"
-            onClick={() => addMetricData(metricDataType)}
-            color="default"
-          />
-        </Col>
-      </Row>
-      {metricData.map((m: MetricMeteredCharge & { localId: string }) => (
-        <div key={m.localId}>
-          <Row key={m.localId} className="my-2">
-            <Col span={5}>
-              <Select
-                style={{ width: 160 }}
-                value={m.metricId}
-                onChange={onMetricIdSelectChange(metricDataType, m.localId)}
-                options={metricsList.map((m) => ({
-                  label: m.metricName,
-                  value: m.id
-                }))}
-              />
-            </Col>
-            <Col span={5}>
-              <Select
-                style={{ width: 160 }}
-                value={m.chargeType}
-                onChange={onChargeTypeSelectChange(metricDataType, m.localId)}
-                options={[
-                  {
-                    label: METRIC_CHARGE_TYPE[MetricChargeType.STANDARD].label,
-                    value: MetricChargeType.STANDARD
-                  },
-                  {
-                    label: METRIC_CHARGE_TYPE[MetricChargeType.GRADUATED].label,
-                    value: MetricChargeType.GRADUATED
-                  }
-                ]}
-              />
-            </Col>
-            <Col span={6}>
-              <InputNumber
-                style={{ width: 120 }}
-                placeholder="Price"
-                prefix={getCurrency()?.Symbol}
-                min={0}
-                value={m.standardAmount}
-                onChange={onMetricFieldChange(
-                  metricDataType,
-                  m.localId,
-                  'standardAmount'
-                )}
-                disabled={m.chargeType == MetricChargeType.GRADUATED}
-              />
-            </Col>
-            <Col span={4}>
-              <InputNumber
-                style={{ width: 120 }}
-                placeholder="Start value"
-                min={0}
-                value={m.standardStartValue}
-                onChange={onMetricFieldChange(
-                  metricDataType,
-                  m.localId,
-                  'standardStartValue'
-                )}
-                disabled={m.chargeType == MetricChargeType.GRADUATED}
-              />
-            </Col>
-            <Col span={2}>
-              <BadgedButton
-                showBadge={m.graduatedAmounts.length > 0}
-                count={m.graduatedAmounts.length}
-              >
-                <Tooltip title="Graduation setup">
-                  <Button
-                    onClick={() =>
-                      setGraduationSetupModalOpen({
-                        metricType: metricDataType,
-                        localId: m.localId
-                      })
-                    }
-                    size="small"
-                    style={{ border: 'none' }}
-                    disabled={m.chargeType == MetricChargeType.STANDARD}
-                    icon={<Graduation />}
-                  />
-                </Tooltip>
-              </BadgedButton>
-            </Col>
-            <Col span={2}>
-              <Button
-                icon={<MinusOutlined />}
-                size="small"
-                style={{ border: 'none' }}
-                onClick={() => removeMetricData(metricDataType, m.localId)}
-              />
-            </Col>
-          </Row>
-          {/* <Row>
-            <Col span={21}>
-              <div className="min-h-10 w-full rounded-md bg-white p-2">
-                graduation setup
-              </div>
-            </Col>
-          </Row> */}
-        </div>
-      ))}
-    </div>
-  )
-}
-
-const GraduationSetupModal = ({
-  data,
-  onCancel,
-  onOK,
-  getCurrency
-}: {
-  data: MetricGraduatedAmount[] | undefined
-  onCancel: () => void
-  onOK: (graduationData: MetricGraduatedAmount[]) => void
-  getCurrency: () => CURRENCY
-}) => {
-  const [graduationData, setGraduationData] = useState<
-    (MetricGraduatedAmount & { localId: string })[]
-  >(
-    data == undefined
-      ? []
-      : data.map((d) => ({ ...d, localId: randomString(8) }))
-  )
-
-  const addGraduationData = () => {
-    setGraduationData(
-      update(graduationData, {
-        $push: [
-          {
-            perAmount: null,
-            startValue: null,
-            endValue: null,
-            flatAmount: null,
-            localId: randomString(8)
-          }
-        ]
-      })
-    )
-  }
-
-  const removeGraduationData = (localId: string) => {
-    const idx = graduationData.findIndex((m) => m.localId == localId)
-    if (idx != -1) {
-      setGraduationData(update(graduationData, { $splice: [[idx, 1]] }))
-    }
-  }
-
-  const onGraduationDataChange =
-    (localId: string, field: keyof MetricGraduatedAmount) =>
-    (val: number | null) => {
-      const idx = graduationData.findIndex((m) => m.localId == localId)
-      if (idx != -1) {
-        setGraduationData(
-          update(graduationData, { [idx]: { [field]: { $set: val } } })
-        )
-      }
-    }
-
-  return (
-    <Modal title="Graduation setup" width={720} open={true} footer={false}>
-      {' '}
-      <Row>
-        <Col span={6}>Per unit</Col>
-        <Col span={6}>Start Value</Col>
-        <Col span={5}>End Value</Col>
-        <Col span={5}>Flat fee</Col>
-        <Col span={2}>
-          <Button
-            icon={<PlusOutlined />}
-            onClick={addGraduationData}
-            style={{ border: 'none' }}
-          />
-        </Col>
-      </Row>
-      {graduationData.map((m) => (
-        <Row key={m.localId} className="my-4">
-          <Col span={6}>
-            <InputNumber
-              style={{ width: 120 }}
-              placeholder="Amount"
-              value={m.perAmount}
-              prefix={getCurrency()?.Symbol}
-              onChange={onGraduationDataChange(m.localId, 'perAmount')}
-            />
-          </Col>
-          <Col span={6}>
-            <InputNumber
-              style={{ width: 120 }}
-              placeholder="Start value"
-              value={m.startValue}
-              onChange={onGraduationDataChange(m.localId, 'startValue')}
-            />
-          </Col>
-          <Col span={5}>
-            <InputNumber
-              style={{ width: 120 }}
-              placeholder="End value"
-              value={m.endValue}
-              onChange={onGraduationDataChange(m.localId, 'endValue')}
-            />
-          </Col>
-          <Col span={5}>
-            <InputNumber
-              style={{ width: 120 }}
-              placeholder="Flat amount"
-              prefix={getCurrency()?.Symbol}
-              value={m.flatAmount}
-              onChange={onGraduationDataChange(m.localId, 'flatAmount')}
-            />
-          </Col>
-          <Col span={2}>
-            <Button
-              style={{ border: 'none' }}
-              icon={<MinusOutlined />}
-              size="small"
-              onClick={() => removeGraduationData(m.localId)}
-            />
-          </Col>
-        </Row>
-      ))}
-      <div className="flex justify-end gap-3">
-        <Button onClick={onCancel}>Cancel</Button>
-        <Button onClick={() => onOK(graduationData)}>OK</Button>
-      </div>
-    </Modal>
-  )
-} //     <Modal title="Graduation setup" width={720} open={true} footer={false}>
-{
-  /* <div className="flex justify-end gap-3">
-        <Button onClick={onCancel}>Cancel</Button>
-        <Button onClick={() => onOK(graduationData)}>OK</Button>
-      </div> */
-}
-
-interface BadgedButtonProps extends PropsWithChildren<object> {
-  showBadge: boolean
-  count: number
-}
-
-const BadgedButton: React.FC<BadgedButtonProps> = ({
-  showBadge,
-  count,
-  children
-}) =>
-  showBadge ? (
-    <Badge count={count} color="gray">
-      {children}
-    </Badge>
-  ) : (
-    children
-  )
