+++ src/components/plan/detail/billableMetric/types.ts
@@ -1,0 +1,7 @@
+import { MetricLimits, MetricMeteredCharge } from '@/shared.types'
+
+export type MetricData = {
+  metricLimits: MetricLimits[]
+  metricMeteredCharge: MetricMeteredCharge[]
+  metricRecurringCharge: MetricMeteredCharge[]
+}
