+++ src/components/plan/index.tsx
@@ -9,6 +9,7 @@ import update from 'immutability-helper'
 import React, { useEffect, useRef, useState } from 'react'
 import { useSearchParams } from 'react-router-dom'
 import DeleteProductModal from './deleteProductModal'
+// import { MetricDataContext } from './detail/metricDataContext'
 import PlanList from './planList'
 import ProductModal from './productModal'
